Fixed Server logic & Tests
All tests pass (except for render which fails due to line-endings)

Previously, when I cleaned up the if/else block for list[bytes], etc I condensed it into the line to compress the body. BUT, that fails when the body is Falsy OR None. We don't want to compress when Falsy, but we also want to convert it to a non-Falsy value (List[bytes]) so I cache the Falsy-ness to get around that. This should emulate previous behaviour while using the new behaviour using strict result rules.